cybercrime/backend/requirements.txt
# FastAPI Core (Python 3.12+ compatible)
fastapi>=0.115.0,<0.116.0
uvicorn[standard]>=0.30.0,<0.31.0
python-multipart>=0.0.9

# Database (Python 3.12+ compatible)
sqlalchemy>=2.0.36,<3.0.0
alembic>=1.13.0,<2.0.0
psycopg2-binary>=2.9.9

# MongoDB
motor>=3.6.0  # Async MongoDB driver
pymongo>=4.10.0  # MongoDB driver

# Auth & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.2.0
python-dotenv>=1.0.1

# Validation (Python 3.12+ compatible - has pre-built wheels)
pydantic>=2.9.0,<3.0.0
pydantic-settings>=2.5.0,<3.0.0
email-validator>=2.2.0

# Crypto
cryptography>=43.0.0

# Utilities
requests>=2.32.0
httpx>=0.27.0
python-dateutil>=2.9.0
aiofiles>=24.1.0
PyYAML>=6.0.1
kafka-python>=2.0.2

# AI (Ollama - Free, Local LLM)
ollama>=0.3.0

# Text-to-Speech (TTS)
TTS>=0.22.0  # Coqui XTTS v2 support
gTTS>=2.5.0  # Google TTS fallback

# Machine Learning dependencies for fraud detection
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Dev tools (optional for production)
pytest>=8.3.0
pytest-asyncio>=0.24.0
black>=24.8.0
flake8>=7.1.0
